# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MyloWare API",
    "description": "Llama Stack-native multi-agent video production platform",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/runs/start": {
      "post": {
        "tags": [
          "Runs"
        ],
        "summary": "Start Run",
        "description": "Start a new workflow run.\n\nCreates a pending run and kicks off workflow execution in the background.\nReturns immediately with the run_id.",
        "operationId": "start_run_v1_runs_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/myloware__api__routes__runs__StartRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/myloware__api__routes__runs__StartRunResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v1/runs/{run_id}": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Get Run",
        "description": "Get run status and artifacts.",
        "operationId": "get_run_v1_runs__run_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/runs/start": {
      "post": {
        "tags": [
          "LangGraph Workflows"
        ],
        "summary": "Start Run",
        "description": "Start a new LangGraph workflow run.\n\nCreates a run in the database and starts the workflow graph.\nReturns immediately if the graph hits an interrupt (HITL or webhook wait).",
        "operationId": "start_run_v2_runs_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/myloware__api__routes__langgraph__StartRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/myloware__api__routes__langgraph__StartRunResponse"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/v2/runs/{run_id}": {
      "get": {
        "tags": [
          "LangGraph Workflows"
        ],
        "summary": "Get Run",
        "description": "Get run status and artifacts (DB projection, not LangGraph state).",
        "operationId": "get_run_v2_runs__run_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetailResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "Machine-readable error code"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail",
            "description": "Human-readable or structured error detail"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunDetailResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "workflow_name": {
            "type": "string",
            "title": "Workflow Name"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "current_step": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Current Step"
          },
          "artifacts": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artifacts"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "workflow_name",
          "status"
        ],
        "title": "RunDetailResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "myloware__api__routes__langgraph__StartRunRequest": {
        "properties": {
          "workflow": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[a-z0-9_]+$",
            "title": "Workflow",
            "description": "Workflow/project name",
            "default": "aismr"
          },
          "brief": {
            "type": "string",
            "maxLength": 10000,
            "title": "Brief",
            "description": "User's brief/prompt"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id",
            "description": "Optional user identifier"
          },
          "telegram_chat_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Chat Id",
            "description": "Optional Telegram chat ID"
          }
        },
        "type": "object",
        "required": [
          "brief"
        ],
        "title": "StartRunRequest",
        "description": "Request to start a new LangGraph workflow run."
      },
      "myloware__api__routes__langgraph__StartRunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "thread_id": {
            "type": "string",
            "title": "Thread Id"
          },
          "state": {
            "additionalProperties": true,
            "type": "object",
            "title": "State"
          },
          "interrupt": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Interrupt"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "thread_id",
          "state"
        ],
        "title": "StartRunResponse",
        "description": "Response from starting a workflow."
      },
      "myloware__api__routes__runs__StartRunRequest": {
        "properties": {
          "project": {
            "type": "string",
            "maxLength": 64,
            "pattern": "^[a-z0-9_]+$",
            "title": "Project",
            "description": "Workflow/project name",
            "default": "aismr"
          },
          "brief": {
            "type": "string",
            "maxLength": 10000,
            "title": "Brief",
            "description": "User's brief/prompt"
          },
          "user_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 128
              },
              {
                "type": "null"
              }
            ],
            "title": "User Id",
            "description": "Optional user identifier"
          },
          "telegram_chat_id": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 64
              },
              {
                "type": "null"
              }
            ],
            "title": "Telegram Chat Id",
            "description": "Optional Telegram chat ID"
          }
        },
        "type": "object",
        "required": [
          "brief"
        ],
        "title": "StartRunRequest",
        "description": "Request to start a new workflow run."
      },
      "myloware__api__routes__runs__StartRunResponse": {
        "properties": {
          "run_id": {
            "type": "string",
            "title": "Run Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "run_id",
          "status"
        ],
        "title": "StartRunResponse"
      }
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}